+++ NyaaTools/mesh/apply_modifiers.py
@@ -1,5 +1,6 @@
 import bpy
 
+from ..avatar.list_avatar_armatures import list_avatar_armatures
 from ..common.deselect_all import deselect_all
 from ..common.selection_add import selection_add
 
@@ -10,14 +11,36 @@ def apply_modifiers(obj):
 
     selection_add(obj)
 
+    avatar_names = list_avatar_armatures()
+
     for modifier in obj.modifiers:
         if (
-            modifier.name == "Armature"
-            or modifier.name.startswith("MToon")
-            or modifier.name.startswith("--")
+            modifier.name.startswith("--")
         ):
+            # Anything with a name starting with "--" is for internal use only and should be removed
+
             bpy.ops.object.modifier_remove(modifier=modifier.name)
+        elif modifier.name.startswith("MToon"):
+            # MToon is a special case where the outline modifier should be removed
+
+            bpy.ops.object.modifier_remove(modifier=modifier.name)
+        elif modifier.type == "ARMATURE":
+            # Remove armature modifiers that target any avatars. These are purely for blender viewing and will be reconfigured by the tool
+            # Otherwise apply it
+
+            if modifier.object == None:
+                bpy.ops.object.modifier_remove(modifier=modifier.name)
+                continue
+
+            armature = modifier.object
+
+            if armature.name in avatar_names:
+                bpy.ops.object.modifier_remove(modifier=modifier.name)
+            else:
+                bpy.ops.object.modifier_apply(modifier=modifier.name)
         else:
+            # Apply the modifier
+            
             bpy.ops.object.modifier_apply(modifier=modifier.name)
 
     try: